Working GUI, SQL, Auth
SQL can now pull databases and their entries. Entries can have any
key/value because they are identified using UUIDs. This prevents entry
mismatch and simplicity. Authentication works by literally connecting
to MySQL and seeing if it works and if there is read/write access,
removing the need for a Fixture “Server” to process authentications.
While this does raise a security risk because the SQL server is wide
open, it is the easiest method for me right now. I may develop a
Fixture server at a later point.